+++ lib/fishstick_dart.dart
@@ -2,6 +2,7 @@ import "client/client.dart";
 
 export "commands/info/ping.dart";
 export "commands/info/invite.dart";
+export "commands/info/info.dart";
 export "commands/partner/premium.dart";
 export "commands/admin/partner.dart";