+++ CHANGELOG.md
@@ -20,3 +20,6 @@
 -   Invite command.
 -   Info command.
 -   Encryption system.
+-   STW missions auto post commands.
+
+## 0.4.0